+++ apps/mobile/src/libs/types/types.ts
@@ -6,6 +6,7 @@ export {
 	type RootNavigationParameterList,
 } from "./navigation/navigation";
 export { type WheelDataItem } from "./wheel/wheel";
+export { type NativeStackNavigationProp } from "@react-navigation/native-stack";
 export {
 	type ImageSourcePropType,
 	type StyleProp,