+++ packages/shared/src/index.ts
@@ -8,6 +8,7 @@ export {
 export {
 	AuthError,
 	HTTPError,
+	OnboardingError,
 	ValidationError,
 } from "./libs/exceptions/exceptions.js";
 export {
@@ -37,6 +38,14 @@ export {
 	AuthApiPath,
 	ConfirmPasswordCustomValidation,
 } from "./modules/auth/auth.js";
+export {
+	type OnboardingAnswerDto,
+	type OnboardingAnswerRequestBodyDto,
+	type OnboardingAnswerRequestDto,
+	type OnboardingAnswerResponseDto,
+	onboardingAnswersValidationSchema,
+	OnboardingApiPath,
+} from "./modules/onboarding/onboarding.js";
 export {
 	type UserDto,
 	type UserGetAllResponseDto,