+++ packages/shared/src/modules/onboarding/libs/types/onboarding-answer-dto.type.ts
@@ -1,0 +1,10 @@
+type OnboardingAnswerDto = {
+	createdAt: string;
+	id: number;
+	label: string;
+	questionId: number;
+	updatedAt: string;
+	userId: number;
+};
+
+export { type OnboardingAnswerDto };